requirements.txt
autopep8==1.6.0
click==8.1.3
configparser==5.2.0
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycodestyle==2.8.0
python-decouple==3.6
python-dotenv==0.20.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.1
